Make ConfigWriter internal lock re-entrant
The ConfigWriter now allows nested locking calls, and acquires
(releases) the config lock from WConfd only the first (last) time.

This is needed for blocks that acquire a lock on the configuration and
use synchronized configuration calls inside. While this should be
avoided in general, it is required for some special circumstances.

Also add a warning that ConfigWriter isn't thread-safe any more.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>