Acquire config lock for RPC calls that check consistency
Since LUClusterVerifyGroup issues 3 RPC calls to get information from
nodes, and during every of these calls, the configuration can be
modified by somebody else, it's necessary to acquire a shared lock on
the configuration.

All the RPC calls are grouped together in a block that holds the lock,
and the tests are only performed afterwards, after the lock is released.

This should be later moved into WConfd, which can prevent configuration
changes internally.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>