Make RAPI opcode attributes use a class and not a tuple
The tuple approach to storing opcode attributes might have worked when
there were four possible modifiers per method, but with aliases this
number has been pushed to five, and with forbidden parameters it will
reach six. This patch refactors it from using a tuple to using a class
with slots, and modifies all the code and tests accordingly.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>