Add a type for the entity owning a lock
While it is unlikely that lock owners will ever be represented by
something different than a job id or, exceptionally, a string, this
type will still improves type checking. Additionally, it will allow
us to make it a separate instance of JSON and thus simplify the
parts of WConfD related to communication.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>