Verify that some progress is made after lock releases
Add a basic sanity check for progress on pending requests. If a
request is pending and all owners it is blocked on release their
locks, there is at least one pending request that can be granted
(the said one). So verify that at least one owner gets notified
of the granting of his request.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>