Provide an extensional representation of LockWaiting
The internal representation of the lock waiting structure contains
some arbitrariness---pending requests are arbitrarily keyed by one
of the lock owners that blocks them. Therefore, LockWaiting is not
an instance of Eq. To allow some meaningful form of comparison
nevertheless (in particular for testing), provide some representation
of LockWaiting where equality on the representation is observable
equivalence.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>