Schedule on jobs where all job dependencies are finished
Jobs may depend on other jobs in the sense that they may only be started
once a given job is finalized. For a job process, however, it is hard to
determine if the status of a different job without a significant overhead.
Therefore, only hand over to execution those jobs where all jobs they depend
on have been finalized.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>