Do not notify the current requester
The current implementation of lock waiting yields as notification
set the list of all owners whose requests could be fulfilled. This
includes the initiating request. While technically correct, the
original requester gets the answer of the request and hence does not
want to have a notification. Therefore, provide external versions
where the original requester is not notified.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>