When forking a job, close all unnecessary file descriptors
This is a bit problematic as there is no portable way how to list all
open file descriptors, and we can't track them all, because they're also
opened by third party libraries such as inotify. Therefore we use
/proc/self/fd and /dev/fd, which should work for all Linux flavors and
most *BSD as well. If both are missing, we don't do anything and just
log a warning.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>