Add verification of compression tool existence
The addition of lzop as a compression option that does not necessarily
exist on a Ganeti-running machine requires checks that lzop exists.
This patch applies a more general solution, and attempts to check for
the existence of a tool by running it with the -h switch on the target
node. A timeout is imposed in case the tool expects input, and missing
tools are handled as well.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Thomas Thrainer <thomasth@google.com>