Add a disk to an instance in config
Implement functions 'AddDisk' and 'AttachInstanceDisk'. The first one
adds a new disk to the config file and the second one attaches a disk to
an instance. A wrapper 'AddInstanceDisk' is provided in order to add a
disk and attach it to an instance at once (because Ganeti doesn't
support dangling disks right now).

Signed-off-by: Ilias Tsitsimpis <iliastsi@grnet.gr>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>