Remove a disk from an instance in config
Implement functions 'RemoveDisk' and 'DetachInstanceDisk'. The first one
removes a disk from the config file and the second one detaches a disk
from an instance. A wrapper 'RemoveInstanceDisk' is provided in order to
detach a disk from an instance and remove it at once (since Ganeti
doesn't support dangling disks yet).

Signed-off-by: Ilias Tsitsimpis <iliastsi@grnet.gr>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>