Implement getDisks in Confd
Add 'ReqInstanceDisks' request type and allow Confd to query for the
disks of an instance. The helper function 'getInstanceDisks' returns the
list of instances on the given node along with their disks and is used
by the function 'addInstNameToLv'.

Update the test's data file 'instance-prim-sec' and rename it to
'instance-disks' as it doesn't return the primary and secondary
instances of a node any more, but it returns the instances and their
disks.

Signed-off-by: Ilias Tsitsimpis <iliastsi@grnet.gr>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>