Lift the Disk objects from the Instances
This patch replaces 'instance.disks' with 'GetInstanceDisks' everywhere
in the codebase. From now on, the function 'GetInstanceDisks' from the
config file has to be used in order to get the disks of an instance.
Also the functions 'AddInstanceDisk'/'RemoveInstanceDisk' have to be
used to add/remove a disk object to/from the config file.

Signed-off-by: Ilias Tsitsimpis <iliastsi@grnet.gr>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>