Forbid the compression tools from being set over RAPI
Although we impose restrictions on what can be input as a compression
tool, someone with RAPI credentials could choose and execute a number
of commands by first setting the custom tools and then executing them.

To prevent this from happening, use the _FORBIDDEN rlib modifier to
forbid tools from being set over RAPI, and verify this in a test.
The QA is also modified to account for the forbidden parameter in
symmetry tests.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Thomas Thrainer <thomasth@google.com>