Add basic test for CPU speed
In this example, the cluster has two nodes and four instances, two with
primary on each of the nodes. The most scarce resource on this cluster
are (virtual) CPUs and the second node has 3 times the CPU speed of the
first one. So distributing the instances 1 and 3 gives a more balanced
cluster.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>