Added support for disk native AIO mode for KVM
This patch adds support for the native aio on KVM
hypervisor.

Basically, It adds a new HV-KVM optional parameter
"disk_aio" that can be set with the followings values:
threads (the default for KVM) or native. If not set, it
uses the KVM default that is "threads". The patch also
updates the man page for gnt-instance.

Signed-off-by: Marco Casavecchia (aka MCM) <marcolinuz@gmail.com>
Reviewed-by: Helga Velroyen <helgav@google.com>