Pass the debug level to forked jobs
When forking off jobs, make them inherit the debug level
of the parent process (i.e., of luxid). In this way, we
can debug jobs in test clusters without cluttering production
logs. We pass the debug level through the environment instead
of the startup dialog with the parent, as we would like to
be able to debug the startup dialog.

While there, fix the call of SetupLogging: the debug parameter
is supposed to be an int.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>