daemon-util: support for daemons loaded on demand
We add a second category of daemons in daemon-util, those that are
started on demand and should not be started by the start_all action, but
should be stopped during service shutdown. This category currently
includes metad and kvmd.

To this end we introduce two more array variables:

 - ON_DEMAND_DAEMONS: currently holds kvmd and metad, which should only
   be stopped during stop-all.

 - ALL_DAEMONS: is the union of DAEMONS and ON_DEMAND_DAEMONS, i.e. the
   list of all known daemons.

We also change the daemon-util test to expect kvmd and metad in the stop
daemon list.

Signed-off-by: Apollon Oikonomopoulos <apoikos@gmail.com>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>