daemon-util: do not restart failed systemd services
daemon-util's check-and-start function is only used by the watcher to
start failed services that should be running on the node. Systemd
supervises and restarts the services as needed; if a service is down, it
is so either because of an unrecoverable error, or because the
administrator manually stopped it. In any case, it should not be blindly
restarted.

Signed-off-by: Apollon Oikonomopoulos <apoikos@gmail.com>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>