kvm: use a dedicated QMP socket for kvmd
The KVM daemon keeps a persistent connection to the instances' QMP
sockets, listening for asynchronous events. As each monitor socket
(either human, or QMP) can handle only one client at a time, this has
the side-effect that QMP cannot be used for regular instance operations.

We fix this by adding a dedicated QMP socket for use by kvmd and
changing kvmd to look only for sockets with the specific extension.

Note that this change means that kvmd will not be able to monitor
already running instances spawned by older versions of Ganeti.

Signed-off-by: Apollon Oikonomopoulos <apoikos@gmail.com>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>