Correctly report user-down instance status
Info queries on user-down instances would fail, because of a check
against instance.hvparams which only contains the instance's non-default
parameters. To fix this, we use GetUpdatedParams, to merge with the cluster
hypervisor parameters.

We also fix a logic error uncovered after the above fix: instances that
were shut down by the user but not marked for user_shutdown, would be
reported as "up" instead of "down".

Signed-off-by: Apollon Oikonomopoulos <apoikos@gmail.com>
Reviewed-by: Jose A. Lopes <jabolopes@google.com>