Make 'access' an optional disk parameter
This patch makes 'access' an optional disk parameter just like
spindles, mode, name, vg and metavg. This option can only be set to
'kernelspace' or 'userspace'. When 'userspace' is used, the instance
will access this disk directly without going through a block device.

If 'access' is not specified, the default value of the node-group
(possibly inherited from the cluster-wide settings) will be used. This
behavior is similar to what 'vg' and 'metavg' parameters do.

Currently this option works only for RBD and Gluster devices, since
these are the only templates that actually support the parameter.

Signed-off-by: Ilias Tsitsimpis <iliastsi@grnet.gr>
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>