Don't synchronize the configuration on DropECReservations
The call doesn't modify the configuration state, only the reservations.
And after moving reservations to WConfd, the reservations state will be
locked just internally by WConfd.

Since this method is called at the end of any job in mcpu, synchronizing
the configuration caused it to be updated in WConfd, even if the job was
read-only.
In particular, "gnt-cluster verify" with multiple node groups often
failed because of this, when verifying configuration consistency.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>