check-man-warnings: use C.UTF-8 and set LC_ALL
check-man-warnings currently partially forces the en_US.UTF-8 locale by
setting LANG. This implicitly assumes that the locale exists, which
might not be the case when building e.g. in chroot environments. If the
locale does not exist, then check-man-warnings fails with the following
message:

 col: Invalid or incomplete multibyte or wide character
 man: command exited with status 1: col -b -p -x

Some distributions (at least Debian and derivatives) ship an embedded
C.UTF-8 locale with full unicode support. We prefer using C.UTF-8 if it
is available and fall back to en_US.UTF-8 otherwise. We also set LC_ALL
to the same locale, because if they remain unset (i.e. "C"), the same
behavior happens.

Signed-off-by: Apollon Oikonomopoulos <apoikos@gmail.com>
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>