Add data type for abstract statistics
Our cluster score is a weighted sum of certain sums and
standard deviations of node characteristics. When placing
a single instance, the cluster score of a big number of
quite similar clusters are computed: that of the original
cluster with one instance added, for each possible placement.
To speed up these computations, add an abstract type of
statistics that allows a potentially more efficient update
of a single value.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>