Modify 'Hypervisor.ListInstances' to exclude user downs
Modify 'Hypervisor.ListInstances' to exclude instances that have been
shutdown by the user.  In other words, 'ListInstances' includes only
instances that are actually running.  This modification requires also
modifying the functions in the 'ganeti.backend' module that start,
stop, and reboot an instance, in order to work properly with user
shutdown.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>