Assume that instance keys might not be present in watcher
Before having the watcher cleanup down instances, the assumption was
that if the instance data was contained in the watcher's state, then
the number of restart attempts would definitely be present.  This is
not true anymore, because it might be the case that if the instance
data is in the watcher's state, it is possible that only one of the
keys is present (or both).  This patch fixes this assumption.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>