Ensure full configuration distribution if MCs change
If a node is removed from the list of master candidates, we want to make
sure that after the operation the configuration will not be distributed
to the node any more. Without requiring full distribution, it's possible
that the node still receives configuration update from a previous call.

Also include the master node(s).
(While this shouldn't change without master failover, if it does, it's
definitely a good idea to flush the configuration.)

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>