Modify how the KVM daemon is started/stopped
This patch modifies how the KVM daemon is started/stopped.  After this
patch, the KVM daemon is always started by default, as it will
terminate itself if necessary, and it is restarted by the watcher.

Also, changes to user shutdown and vm capable attributes will cause
the KVM daemon instances on the corresponding nodes to be
started/stopped accordingly.  This is achieved through the helper
function 'EnsureKvmdOnNodes'.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>