Make mcpu use wconfd's asynchronous requests
Instead of only polling make mcpu ask wconfd to hand out the
locks as soon as they become available. For the time being,
we still poll to find out if the request was granted. Nevertheless,
this will ensure that locks are handed out in the intended order,
in particular honoring job priorities.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>