Add a constant block size for the 'dd' command
Create a new constant variable named 'ddBlockSize' that holds the
default value for the block size option of the 'dd' command, i.e., the
'bs' option. In Ganeti the block size is always 1MiB to match its units,
so we create a new constant variable for it, instead of using hardcoded
values.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>