Support disk hotplug with userspace access
Before any hotplug action we assemble the block device. Currently
call_blockdev_assemble() returns the link_name as calculated by
_SymlinkBlockDev().

With userspace support we have to return the drive_uri as calculated
by _CalculateDeviceURI() as well, in order for the drive_add monitor
command to be able to use it.

Additionally with this patch the runtime files are properly updated
to include the drive uri as well, and thus, upon instance migration,
the target process will be started with the correct drive options.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>