Simplify type for RuntimeData
In the run-time data for lock queries, the lock itself is never
used, as the name is already provided by the framework from the
list of all locks. Moreover, as everything remaining are lists
which have a suitable zero element, we can make them unconditional.
As a side-effect of this change, we also get correct information
for not-owned locks that have pending requests.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>