Introduce the bdev 'Import' and 'Export' methods
Extend the block device abstract class with two new methods, named
'Import' and 'Export'. Those methods will be responsible for building
the commands which will be used for the disk data copy when we convert
the disk template of the instance. The 'Export' method will build the
command which will export the data from the source disk device, while
the 'Import' will import the data to the newly created target device.

By default the data import and export will be done using the Unix 'dd'
command. All the inherited classes will use the base functionality
unless there is a faster way to convert to. In that case the underlying
block device can override those methods with its specific utility.

Note, that those two methods will not perform the actual data copy
between the source and target devices. Their output will be first
concatenated using a pipe ("|"), and then executed by the caller method
in the backend level.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>