Introduce new RPC call 'blockdev_convert'
Introduce a new RPC call named 'blockdev_convert'. This method uses the
'backend.BlockdevConvert' function which takes as arguments the source
and target disks between which we will perform the actual data copy of
their contents.

This method first gets the export and import commands from the source
and target devices respectively, and then concatenates them to a unified
command using a pipe ("|"), following a similar approach with the impexp
daemon. Finally, the unified command is executed by this method.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>