Add unit tests for blockdev Import/Export methods
Extend the unit test 'ganeti.storage.bdev_unittest.py' to support the
new 'Import' and 'Export' functionality in the block device abstract
class.

This patch tests whether the 'Import' and 'Export' methods return the
desirable output command. It covers the 'RADOSBlockDevice' volumes that
override the default 'dd' behavior with the 'rbd' utility, and also the
'PersistentBlockDevice' template that does not comply completely with
the default case. The rest of the block devices inherit the default 'dd'
functionality from the base class. So, to avoid repeating tests for the
same behavior for all those classes (FileStorage, Gluster, DRBD) we
write tests only for the 'LogicalVolume' class.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>