Drop active flag in the worker pool
Given that now each process only runs a single job, there is no need
anymore to keep track about the worker pool being active. Hence drop
the corresponding code.

The only active use of the SetActive functionality is in the tests;
these tests, however, test priority handling in the worker pool,
which is not that important for single-worker worker pools anyway.
So we just remove the tests.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>