Remove the 'disks_info' attribute from 'Instance'
This attribute is used only in the backend and added automatically
during RPC calls. Therefore, it is just declared in Instance,
but omitted during serialization, and it's deserialized only if
it is really present.

This is the same approach as with Instance.secondary_nodes, but in
addition, disks_info's deserialization must be customized, if present.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>