Export logical_id info to hooks env
..and let utils compute the hooks environment related to disks.

Until now, the only exported information regarding disks in hooks
context was their size, mode, name and uuid.

With this patch, depending on the disk template, we export the info
contained in the logical_id of the disk (e.g. driver, id, provider,
drbd's minors, port, etc.) along with the template name as well.

Introduce a new helper method, BuildDiskEnv(), that takes
either a disk object or a dict with IDISK_PARAMS returned by
ComputeDisks(), and creates the proper hooks' environment for
the corresponding disk.

In case the argument is a Disk object then it updates the
environment with the driver and id info extracted from the
logical_id and based on its template (dev_type).

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>
Reviewed-by: Helga Velroyen <helgav@google.com>