Shutdown instance without checking
Checking if an instance is running and then shutting it down consumes
min 1 RPC and max 2 RPCs.  Given that shutting down an instance that
is already shutdown is an idempotent operation, we can simply shutdown
the instance without checking, and thus reduce the max RPCs to 1.
This change also has the additional advantage that the backend already
knows how to handle with user shutdown instances, so we don't have to
duplicate that logic to the cmdlib as well.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>