Fix 'IsInstanceRunning' to properly detect user down insts
The previous merge 3c8877d0a695e52d12864571ead5b878cda1b4b8 changed
'ListInstances' not to return user down instances.  Therefore, we must
change 'IsInstanceRunning' not to rely on 'call_instance_list' but
instead on 'call_instance_info'.  Also, because of the change in the
previous patch, the parameter 'check_user_shutdown' became obsolete so
it can also be removed.

The only call to 'IsInstanceRunning' that is affected by the parameter
'check_user_shutdown' is in 'CheckInstanceState'.  However, this
change should not change the semantics of 'CheckInstanceState' in a
way that will break the LUs.

Signed-off-by: Jose A. Lopes <jabolopes@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>