Fix LUNetworkConnect wrt vlan
A network is currently connected to a nodegroup with a mode and a
link. Those netparams are inherited as nicparams when a NIC is
attached to a network and filled with the cluster wide defaults.

Since vlan was added to NIC parameters, netparams should be updated
to include it too otherwise NICs attached to a network can be found
with mode=bridged, link=some bridge and the cluster wide vlan.

This patch fixes LUNetworkConnect to add an empty string vlan to
netparams along with the user provided mode and link. Additionally
it adds an empty vlan to existing netparams via
objects.NodeGroup.UpgradeConfig().

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>