Make sure Ssconf is distributed if it affects kvmd's state
In particular, if enabled_user_shutdown or enabled_hypervisors flags
change, or if one of the nodes changes its vm_capable status.
This means that on some nodes kvmd will need to be started or stopped,
which in turn means that ssconf needs to be synchronized on the nodes.

While at it, simplify 'needsFullDist' slightly.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>