When definitely requesting locks, do so one by one
The strategy for obtaining locks is that we first, for a given time,
only try to get locks when all locks are available in order not to
hold unused resources. After a certain time, however, we need to make
sure that we eventually get all the locks we need. This is done by
holding all locks obtained so far. In order for this strategy to work,
however, also the locks of every level need to be obtained sequentially.
Otherwise, if two locks at a level are requested, other processes might still
starve us.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>